-applied permissions in view to price type

+++ Shopper/Services/Implementations/ReportService.cs
@@ -72,10 +72,10 @@ namespace Shopper.Services.Implementations
             return new DashboardSummary
             {
                 SummaryType = SummaryType.Today,
-                ItemsSold = (uint) itemsSold,
-                Sales = (ulong) totalSales,
-                Expenditure = (ulong) expenditure,
-                Profit = (ulong) (profit - expenditure)
+                ItemsSold = itemsSold,
+                Sales = totalSales,
+                Expenditure = expenditure,
+                Profit = profit - expenditure
             };
         }
 
@@ -96,10 +96,10 @@ namespace Shopper.Services.Implementations
             return new DashboardSummary
             {
                 SummaryType = SummaryType.Week,
-                ItemsSold = (uint) itemsSold,
-                Sales = (ulong) totalSales,
-                Expenditure = (ulong) expenditure,
-                Profit = (ulong) (profit - expenditure)
+                ItemsSold = itemsSold,
+                Sales = totalSales,
+                Expenditure = expenditure,
+                Profit = (profit - expenditure)
             };
         }
 
@@ -120,10 +120,10 @@ namespace Shopper.Services.Implementations
             return new DashboardSummary
             {
                 SummaryType = SummaryType.Month,
-                ItemsSold = (uint) itemsSold,
-                Sales = (ulong) totalSales,
-                Expenditure = (ulong) expenditure,
-                Profit = (ulong) (profit - expenditure)
+                ItemsSold = itemsSold,
+                Sales = totalSales,
+                Expenditure = expenditure,
+                Profit = (profit - expenditure)
             };
         }
 
@@ -144,10 +144,10 @@ namespace Shopper.Services.Implementations
             return new DashboardSummary
             {
                 SummaryType = SummaryType.Year,
-                ItemsSold = (uint) itemsSold,
-                Sales = (ulong) totalSales,
-                Expenditure = (ulong) expenditure,
-                Profit = (ulong) (profit - expenditure)
+                ItemsSold = itemsSold,
+                Sales = totalSales,
+                Expenditure = expenditure,
+                Profit = (profit - expenditure)
             };
         }
     }
